rename sessionRecording to session-recording
