+++ frontend/src/scenes/persons/Person.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import { Row, Tabs, Col, Card, Skeleton, Tag, Dropdown, Menu, Button, Popconfirm } from 'antd'
 import { SessionsView } from '../sessions/SessionsView'
 import { EventsTable } from 'scenes/events'
-import { SessionRecordingsTable } from 'scenes/sessionRecordings/SessionRecordingsTable'
+import { SessionRecordingsTable } from 'scenes/session-recordings/SessionRecordingsTable'
 import { useActions, useValues } from 'kea'
 import { personsLogic } from './personsLogic'
 import { PersonHeader } from './PersonHeader'
