+++ frontend/src/scenes/session-recordings/DurationFilter.tsx
