+++ frontend/src/scenes/session-recordings/SessionRecordingTable.scss
