+++ frontend/src/scenes/session-recordings/SessionRecordingsTable.tsx
