+++ frontend/src/scenes/session-recordings/__stories__/Recordings.stories.tsx
