+++ frontend/src/scenes/session-recordings/__stories__/recordings.json
