+++ frontend/src/scenes/session-recordings/player/PlayerController.tsx
