+++ frontend/src/scenes/session-recordings/player/PlayerEvents.tsx
