+++ frontend/src/scenes/session-recordings/player/PlayerFrame.tsx
