+++ frontend/src/scenes/session-recordings/player/sessionRecordingPlayerLogic.ts
