+++ frontend/src/scenes/session-recordings/sessionRecordingsTableLogic.test.ts
